projeweb kopyası/requirements.txt
Django==5.2
python-dotenv==1.0.1
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
supabase==2.3.4
python-jose[cryptography]==3.3.0
httpx==0.26.0 